Added a build tag specific to systemd and updated the documentation
